pyproject.toml
[project]
name = "spark-optimisation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dateutils>=0.6.12",
    "pyspark>=4.0.0",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
    "spark-optimisation",
]

[tool.ruff]
line-length = 127

[tool.uv.sources]
spark-optimisation = { workspace = true }
